backend/requirements.txt
fastapi
uvicorn[standard]
requests
numpy
duckduckgo-search
python-dotenv
pydantic